+++ lib/src/core/animation_strategy.dart
@@ -3,7 +3,6 @@
 *  Licensed under the BSD-style license. See LICENSE in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';